rename FlightLog to FlightEntry
I planned my API around working with FlightLogs as distinct from their entries.

My groupmates stated that they wanted to work with entries, not logs,
so I removed the conception of a log
from the code base.

I kept the endpoint and DB table naems as FlightLog
because that's what I stated in the preliminary API docs I
wrote, but, honestly, it's a lot cleaner to revert back
to the entry name for everything.